history: fix filtering by expressions
It is possible to sequentially filter run results QuerySet by multiple
expressions. And since filtering by expression includes methods that
cannot be applied to a union, it is necessary to return not a union,
but a new identical QuerySet as the result of filtering by expression.
Then filtering by expressions is not necessarily the last thing
to apply.

Signed-off-by: Natalia Rybchenko <natalia.rybchenko@oktetlabs.ru>