importruns: fix measurement results processing
As a result of changes in the measurement results processing mechanism,
the order of processing entries was modified, which in turn affects the
serial number assigned to the measurement result. Therefore, during
forced import, the same results may have different serial numbers and be
treated as distinct, leading to object duplication. To prevent this, the
serial number must be excluded from the check for measurement result
existence in the database. This is correct because measurement result
series are now stored as separate objects, and seriality of
MeasurementResult objects is no longer supported.

Issue: #109
Signed-off-by: Natalia Rybchenko <natalia.rybchenko@oktetlabs.ru>